improved architecture of ContactInfo react compoents
