+++ backend/talent/serializers.py
@@ -29,8 +29,8 @@ class TalentSerializer(serializers.ModelSerializer):
     talent_medicals = TalentMedicalSerializer(many=True, read_only=True)
     talent_availabilities = TalentAvailabilitySerializerWithoutTalentID(many=True, read_only=True)
     talent_ratings = TalentRatingSerializer(many=True, read_only=True)
-    average_rating = serializers.FloatField(source='get_average_rating')
-    profile_status = serializers.JSONField(source='get_profile_status')
+    average_rating = serializers.FloatField(source='get_average_rating', read_only=True)
+    profile_status = serializers.JSONField(source='get_profile_status', read_only=True)
     # is_completed_contact_info = serializers.BooleanField(source='get_is_completed_contact_info')
     # is_completed_nationality = serializers.BooleanField(source='get_is_completed_nationality')
     # is_completed_language = serializers.BooleanField(source='get_is_completed_language')
